+++ staticsite/archetypes.py
@@ -45,7 +45,7 @@ class Archetypes:
                 if f.startswith("."):
                     continue
                 relpath = os.path.relpath(os.path.join(root, f), self.root)
-                for feature_name, feature in self.site.features.items():
+                for feature in self.site.features.ordered():
                     a = feature.try_load_archetype(self, relpath, name)
                     if a is not None:
                         return a